Move eval_ fns out from math_ops and strings ...
and go over docstring tags for these.
