+++ mathics/builtin/matrices/constrmatrix.py
@@ -8,7 +8,6 @@ Methods for constructing Matrices.
 
 from mathics.builtin.base import Builtin
 from mathics.core.atoms import Integer0
-from mathics.core.expression import Expression
 from mathics.core.list import ListExpression
