+++ mathics/data/.gitignore
@@ -1,1 +1,2 @@
 /doc_tex_data.pcl
+/op-tables.json
